pyproject.toml
[tool.poetry]
name = "Wadden-Sea-backend"
version = "0.1.0"
description = "Overall package"
authors = ["Mostafa Farrag <mostafa.farrag@deltares.nl>"]
license = "Proprietary"
readme = "README.md"
keywords = ["llm", "generativeai", "chatbot", "rag", "natural language processing"]
repository = "https://github.com/Deltares-research/Wadden-Sea-backend"

packages = [{ include = "wadden_sea", from = "src"}]


[tool.poetry.urls]

[tool.poetry.dependencies]
python = ">=3.12,<4"
python-dotenv = "^1.0.1"
vfn-rag = {git = "https://github.com/Deltares-research/vfn-rag", rev = "0.2.0"}
azure-keyvault-secrets = "^4.10.0"


[tool.poetry.group.analysis.dependencies]
jupyterlab= "^4.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
xdoctest = {extras = ["colors"], version = "^1.1.0"}
coverage = {extras = ["toml"], version = "^7.6.2"}



[tool.coverage.report]
show_missing = true
fail_under = 10


[tool.pytest.ini_options]
addopts = [ "--cov"]

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
